Sounds:
Derrick S said:

Hello Victor,
Here are all the sounds for the UI. The majority of these sounds are single audio effects and some may run a few extra seconds. The sound suppose to reflect the flow of the animation on the UI when engage.

Here are the sounds and MP3 files.
1. This sound effect happen when you click on the game screenshots on the home UI. It's called (robot_blip)
2. This sound is when a user click on any of the setting, login, log off, or register buttons. It's called Ting_Popup_pixels
3. This sound is used when a gamer scroll through the games on the home UI. It's called Thrown_shuriken
4.When users click on the question mark in the create form. The sound is called Click2_Sebastian
5.When users click the arrow for the next form in the create tab. The sound is called waxing_eyebrows_OTON
6. Create Game Sound. This sound is engage when the user click on the Create game or Edit Game tabs. It's called Electrical_Sweep
7.Play, save, play with friends. This is after the game has been selected. The sound is called Robot_blip2
8.This sound is when users select items in the create tab selection. Example when a users select Action for game genre this effect will be used each time when making a selection for the create forms. Button Click On

If i missed a sound please let me know. but this should be all the sounds.

Thanks,

Derrick